CaptainsChair/server/requirements.txt
flask==1.1.2
flask_restful==0.3.8
requests==2.24.0
python-dotenv==0.14.0
tweepy==3.9.0